requirements.txt
gunicorn==20.0.4
# gevent==20.6.2

Pillow==7.1.2

pathlib==1.0.1

django==3.0.7
djangorestframework==3.11.0
django-redis==4.12.1
django-cors-headers==3.4.0

psycopg2-binary==2.8.5